requirements.txt
websocket-client
python-binance
pandas
matplotlib
cryptoxlib-aio